# public/swagger.json
{
	"components": {
...
		"schemas": {
			"ResponseData": {
				"properties": {
					"status": {
						"type": "number",
						"enum": [0, 1],
						"description": "1 = succeed | 0 = failed"
					},
					"data": {},
					"messages": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Error/warning messages"
					}
				},
				"required": ["status", "data", "messages"],
				"type": "object",
				"additionalProperties": false
			},
			"BuildStatus": {
				"type": "string",
				"enum": ["start", "building", "failed", "success"]
			},
			"ObjectID": {
				"description": "Create a new ObjectID instance.",
				"properties": {
					"generationTime": {
						"type": "number",
						"format": "double",
						"description": "The generation time of this ObjectId instance."
					}
				},
				"required": ["generationTime"],
				"type": "object",
				"additionalProperties": false
			},
			"RegistryProviderType": {
				"type": "string",
				"enum": ["gcloud", "digitalocean", "dockerhub"]
			},
			"ProviderInfo": {
				"properties": {
					"name": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"access_token": {
						"type": "string"
					}
				},
				"required": ["name"],
				"type": "object",
				"additionalProperties": false
			},
			"AccessTokenInfo": {
				"properties": {
					"access_token": {
						"type": "string"
					},
					"expiredTimestamp": {
						"type": "number",
						"format": "double"
					},
					"expiredDate": {
						"type": "string",
						"format": "date-time"
					},
					"expiredDateGTM7": {
						"type": "string"
					}
				},
				"required": ["access_token", "expiredTimestamp", "expiredDate", "expiredDateGTM7"],
				"type": "object",
				"additionalProperties": false
			},
...
			"User": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string",
						"description": "User name"
					},
					"username": {
						"type": "string",
						"description": "Unique username of a user\nThis equavilent with \"slug\""
					},
					"type": {
						"type": "string",
						"enum": ["user", "service_account"],
						"description": "Service Account is User!"
					},
					"email": {
						"type": "string",
						"description": "User email address"
					},
					"verified": {
						"type": "boolean",
						"description": "Is this user's email or phone verified?"
					},
					"image": {
						"type": "string",
						"description": "User profile picture URL"
					},
					"providers": {
						"items": {
							"$ref": "#/components/schemas/ProviderInfo"
						},
						"type": "array",
						"description": "List of Cloud Providers which this user can access to"
					},
					"password": {
						"type": "string",
						"description": "User password (hashed)"
					},
					"token": {
						"$ref": "#/components/schemas/AccessTokenInfo",
						"description": "User token"
					},
					"roles": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/ObjectID"
								},
								{
									"$ref": "#/components/schemas/Role"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array",
						"description": "User's roles\nTODO: Thêm \"workspace\" cho \"roles\" vì mỗi workspace sẽ có roles khác nhau, trong khi mỗi user có thể nằm trong nhiều workspace"
					},
					"teams": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/ObjectID"
								},
								{
									"$ref": "#/components/schemas/Team"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array",
						"description": "User's team IDs which this user is a member"
					},
					"workspaces": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/ObjectID"
								},
								{
									"$ref": "#/components/schemas/Workspace"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array",
						"description": "List of workspace IDs which this user is a member"
					},
					"activeWorkspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "Active workspace of a user"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Project": {
				"description": "Projects",
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"apiKey": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"lastUpdatedBy": {
						"type": "string"
					},
					"latestBuild": {
						"type": "string"
					},
					"appSlugs": {
						"type": "string"
					},
					"apps": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/ObjectID"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/App"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						],
						"description": "List of App IDs"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the workspace"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"WorkspaceApiAccessToken": {
				"properties": {
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/Role"
						},
						"type": "array"
					}
				},
				"required": ["name", "token"],
				"type": "object",
				"additionalProperties": false
			},
			"Workspace": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Workspace slug: auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string",
						"description": "Workspace name"
					},
					"public": {
						"type": "boolean",
						"description": "Is this a `public` workspace that anyone can join, otherwise it's a `private` workspace that needs approval on joining request."
					},
					"image": {
						"type": "string",
						"description": "Workspace profile picture"
					},
					"domain": {
						"type": "string",
						"description": "Workspace domain name"
					},
					"apiAccessTokens": {
						"items": {
							"$ref": "#/components/schemas/WorkspaceApiAccessToken"
						},
						"type": "array",
						"description": "List of this workspace's API Access Token"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"App": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"lastUpdatedBy": {
						"type": "string"
					},
					"git": {
						"properties": {
							"repoSSH": {
								"type": "string"
							},
							"repoURL": {
								"type": "string"
							},
							"provider": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"framework": {
						"properties": {
							"repoSSH": {
								"type": "string"
							},
							"repoURL": {
								"type": "string"
							},
							"slug": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"environment": {
						"properties": {},
						"additionalProperties": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/DeployEnvironment"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "object",
						"deprecated": true
					},
					"deployEnvironment": {
						"properties": {},
						"additionalProperties": {
							"$ref": "#/components/schemas/DeployEnvironment"
						},
						"type": "object"
					},
					"latestBuild": {
						"type": "string"
					},
					"projectSlug": {
						"type": "string"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "Owner ID of the app"
					},
					"project": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Project"
							},
							{
								"type": "string"
							}
						],
						"description": "Project ID of the app"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "Workspace ID of the app"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ContainerRegistry": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"isVerified": {
						"type": "boolean"
					},
					"host": {
						"type": "string",
						"description": "The host (domain) of your container registry which you are using."
					},
					"imageBaseURL": {
						"type": "string",
						"description": "Base URL of the image, usually is the registry host URI combines with something else.\n- This will be used to combine with your project/app image path."
					},
					"provider": {
						"$ref": "#/components/schemas/RegistryProviderType",
						"description": "Provider's \"shortName\""
					},
					"serviceAccount": {
						"type": "string",
						"description": "Content of the Service Account credentials ti access services on this cloud provider\n- Apply for: Google Cloud, AWS,...\n- For example: Kubernetes Clusters, Single Sign-On,..."
					},
					"apiAccessToken": {
						"type": "string",
						"description": "Content of the API access token to use services on this cloud provider\n- Apply for: Digital Ocean"
					},
					"imagePullSecret": {
						"properties": {
							"value": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the workspace"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Build": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"tag": {
						"type": "string",
						"description": "Image tag is also \"buildNumber\""
					},
					"startTime": {
						"type": "string",
						"format": "date-time",
						"description": "Build start time"
					},
					"endTime": {
						"type": "string",
						"format": "date-time",
						"description": "Build end time"
					},
					"duration": {
						"type": "number",
						"format": "double",
						"description": "Build duration in miliseconds"
					},
					"env": {
						"type": "string",
						"description": "Build for which deploy environment\n- **[OPTIONAL] SHOULD NOT rely on this!**\n- A build should be able to used for any deploy environments."
					},
					"branch": {
						"type": "string",
						"description": "Build from which git branch"
					},
					"cliVersion": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/BuildStatus"
					},
					"projectSlug": {
						"type": "string"
					},
					"appSlug": {
						"type": "string"
					},
					"logs": {
						"type": "string"
					},
					"registry": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/ContainerRegistry"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the container registry"
					},
					"app": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/App"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the app"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					},
					"project": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Project"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the project"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the workspace"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Omit_Build.keyofHiddenBodyKeys_": {
				"$ref": "#/components/schemas/Pick_Build.Exclude_keyofBuild.keyofHiddenBodyKeys__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"BuildData": {
				"$ref": "#/components/schemas/Omit_Build.keyofHiddenBodyKeys_"
			},
...
			"StartBuildParams": {
				"properties": {
					"isDebugging": {
						"type": "boolean",
						"default": "false"
					},
					"cliVersion": {
						"type": "string",
						"description": "Diginext CLI version of client user"
					},
					"buildDir": {
						"type": "string",
						"description": "Path to the source code directory\n* [OPTIONAL] Only apply for CLI command, has no effects on API call"
					},
					"env": {
						"type": "string",
						"description": "Select the deploy environment to build image, in this case, this info is using for selecting \"Dockerfile\"\nof specific deploy environment only, for example: \"Dockerfile.dev\" or \"Dockerfile.prod\",\nif you don't specify \"env\", a default \"Dockerfile\" will be selected.\n- **[OPTIONAL] SHOULD NOT rely on this!**\n- A build should be able to used for any deploy environments."
					},
					"registrySlug": {
						"type": "string",
						"description": "Slug of the Container Registry"
					},
					"userId": {
						"type": "string",
						"description": "ID of the author"
					},
					"gitBranch": {
						"type": "string",
						"description": "Select a git branch to pull source code & build"
					},
					"buildNumber": {
						"type": "string",
						"description": "Build number is also an container image's tag"
					},
					"appSlug": {
						"type": "string",
						"description": "App's slug"
					}
				},
				"required": ["registrySlug", "userId", "gitBranch", "buildNumber", "appSlug"],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "API_ACCESS_TOKEN",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "@topgroup/diginext",
		"version": "2.6.5",
		"description": "A BUILD SERVER & CLI to deploy apps to any Kubernetes clusters.",
		"license": {
			"name": "GPL"
		},
		"contact": {
			"name": "TOP GROUP (a.k.a Digitop)",
			"email": "dev@wearetopgroup.com",
			"url": "https://github.com/digitopvn/diginext"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/build": {
			"get": {
				"operationId": "Read",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/Build"
														},
														{
															"items": {
																"$ref": "#/components/schemas/Build"
															},
															"type": "array"
														}
													]
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Build"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Find one item by `{ObjectID}`",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Mark this request as search (return the similar results based on the filter query params)",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"default": "true",
							"type": "boolean"
						}
					},
					{
						"description": "If `true`, return the excel binary file to download.",
						"in": "query",
						"name": "download",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/Build"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Build"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BuildData"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/Build"
													},
													"type": "array"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Build"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BuildData"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"n": {
															"type": "number",
															"format": "double"
														},
														"ok": {
															"type": "number",
															"format": "double"
														}
													},
													"type": "object"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Build"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Delete one item by `{ObjectID}`",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Delete one item by `{slug}`",
						"in": "query",
						"name": "slug",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/build/logs": {
			"get": {
				"operationId": "GetLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"type": "string"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Build"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/build/start": {
			"post": {
				"operationId": "StartBuild",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"description": "Create a new {Build} instance, then start building container image.",
				"tags": ["Build"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StartBuildParams"
							}
						}
					}
				}
			}
		},
		"/build/stop": {
			"patch": {
				"operationId": "StopBuild",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/Build"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Build"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"slug": {
										"type": "string"
									}
								},
								"required": ["slug"],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}